Tips move from a decimal stored as SQLite TEXT to whole integer cents, with a
hand-written migration that converts the values in place rather than the
drop-and-add EF would have scaffolded, and $00.00 presentation and entry.

In its own window KHost stops behaving like a browser: no page context menu
outside text fields, no reload or back shortcuts, and no developer tools, all
gated on a Release build.

Panel resizing works again — it had been dead whenever the page first rendered
with no singer selected, which is always when the venue clears its queue on
close.

+++ src/KHost.Domain/Services/TipsService.cs
@@ -27,6 +27,6 @@ public class TipsService : BaseRepositoryService<Tip, ITipsRepository>, ITipsSer
     public Task<IReadOnlyList<Tip>> GetByUserIdAsync(Guid userId)
         => Repository.GetByUserIdAsync(userId);
 
-    public Task<decimal> GetTotalByUserIdAsync(Guid userId, DateTime? from = null, DateTime? to = null)
+    public Task<int> GetTotalByUserIdAsync(Guid userId, DateTime? from = null, DateTime? to = null)
         => Repository.GetTotalByUserIdAsync(userId, from, to);
 }
